+++ mobile/lib/backend.dart
@@ -78,6 +78,8 @@ Future<void> _startBackend({seedDir, fcmToken}) async {
   try {
     await rust.api.runInFlutter(seedDir: seedDir, fcmToken: fcmToken);
   } catch (e) {
+    // TODO(holzeis): We should add a more gracefull handling of an error during startup here. At
+    // the moment the user will have no idea what happened and is not able to share his logs.
     logger.e("Launching the app failed $e");
     await Future.delayed(const Duration(seconds: 5));
     exit(-1);